change prisma option  (#321)
