+++ prisma/schema.prisma
@@ -14,7 +14,7 @@ generator fabbrica {
   provider    = "prisma-fabbrica"
   output      = "../src/lib/services/.fabbrica"
   tsconfig    = "../tsconfig.json"
-  noTranspile = false
+  noTranspile = true
 }
 
 datasource db {
